@@ -19,8 +19,7 @@
     "electron-dev": "NODE_ENV=development ELECTRON_START_URL=http://localhost:3000 electron electron/electron.js",
     "react-start": "BROWSER=none; NODE_ENV=development react-scripts start",
     "react-build": "react-scripts build",
-    "test": "mocha",
-    "cover": "rm -Rf coverage/* && istanbul cover _mocha -- -R spec",
+    "test": "jest",
     "lint": "eslint '*.js' 'electron/**/*.js' 'api/**/*.js' 'src/**/*.js' 'test/**/*.js'",
     "report": "coveralls < ./coverage/lcov.info",
     "prebuild": "rm -rf dist/ && rm -rf build/",
@@ -33,27 +32,25 @@
   "proxy": "http://localhost:2600/",
   "main": "electron/electron.js",
   "dependencies": {
-    "aws-sdk": "^2.1094.0",
-    "body-parser": "^1.19.2",
+    "@aws-sdk/client-sts": "^3.282.0",
+    "@node-saml/passport-saml": "^4.0.2",
+    "@xmldom/xmldom": "^0.8.7",
+    "body-parser": "^1.20.2",
     "cookie-parser": "^1.4.6",
-    "errorhandler": "^1.5.0",
-    "express": "^4.17.3",
-    "express-session": "^1.17.2",
+    "express": "^4.18.2",
+    "express-session": "^1.17.3",
     "ini": "^1.3.8",
-    "lodash.isplainobject": "^4.0.6",
-    "mkdirp": "^0.5.5",
     "morgan": "^1.10.0",
-    "passport": "^0.4.1",
-    "passport-saml": "^3.1.0",
-    "react": "^16.14.0",
-    "react-dom": "^16.14.0",
-    "uuid": "^3.4.0",
-    "xmldom": "^0.6.0",
+    "passport": "^0.6.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "uuid": "^9.0.0",
     "xpath.js": "^1.1.0"
   },
   "devDependencies": {
     "@babel/core": "^7.21.0",
     "@babel/eslint-parser": "^7.19.1",
+    "@babel/preset-env": "^7.21.4",
     "@babel/preset-react": "^7.18.6",
     "@fortawesome/fontawesome-free": "^6.3.0",
     "@fortawesome/fontawesome-svg-core": "^6.3.0",
@@ -61,24 +58,24 @@
     "@fortawesome/free-solid-svg-icons": "^6.3.0",
     "@fortawesome/react-fontawesome": "^0.2.0",
     "@popperjs/core": "^2.11.6",
-    "axios": "^0.21.2",
+    "axios": "^1.3.4",
+    "babel-jest": "^29.5.0",
     "bootstrap": "^5.2.3",
     "coveralls": "^3.1.1",
-    "electron": "^23.1.1",
+    "electron": "^23.1.2",
     "electron-packager": "^17.1.1",
     "electron-rebuild": "^3.2.9",
     "eslint": "^8.35.0",
     "eslint-config-airbnb": "^19.0.4",
     "eslint-plugin-import": "^2.27.5",
+    "eslint-plugin-jest": "^27.2.1",
     "eslint-plugin-jsx-a11y": "^6.7.1",
     "eslint-plugin-node": "^11.1.0",
     "eslint-plugin-react": "^7.32.2",
     "eslint-plugin-react-hooks": "^4.6.0",
-    "history": "^4.10.1",
-    "istanbul": "^0.4.5",
-    "mocha": "^5.2.0",
-    "mocha-lcov-reporter": "^1.3.0",
-    "prismjs": "^1.27.0",
+    "history": "^5.3.0",
+    "jest": "^29.5.0",
+    "prismjs": "^1.29.0",
     "prop-types": "^15.8.1",
     "react-is": "^18.2.0",
     "react-router": "^6.8.2",
